+++ src/main/java/com/couponmoa/backend/couponmoauser/domain/user/controller/UserControllerV1.java
@@ -1,20 +1,17 @@
-package com.couponmoa.backend.couponmoauser.domain.user.controller.v1;
+package com.couponmoa.backend.couponmoauser.domain.user.controller;
 
 import com.couponmoa.backend.couponmoauser.common.dto.ApiResponse;
 import com.couponmoa.backend.couponmoauser.domain.user.dto.request.UserDeleteRequest;
 import com.couponmoa.backend.couponmoauser.domain.user.dto.request.UserUpdatePasswordRequest;
 import com.couponmoa.backend.couponmoauser.domain.user.dto.request.UserUpdateRequest;
 import com.couponmoa.backend.couponmoauser.domain.user.dto.response.UserResponse;
-import com.couponmoa.backend.couponmoauser.domain.user.service.v1.UserServiceV1;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
+import com.couponmoa.backend.couponmoauser.domain.user.service.UserServiceV1;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-@Tag(name = "사용자 관리 API", description = "사용자 관리 API")
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/v1/users")
@@ -22,14 +19,12 @@ public class UserControllerV1 {
 
     private final UserServiceV1 userServiceV1;
 
-    @Operation(summary = "본인 정보 조회", description = "본인의 사용자 정보를 확인함")
     @GetMapping
     public ResponseEntity<ApiResponse<UserResponse>> findUser(@RequestHeader("X-User-Id") Long userId) {
         UserResponse userResponse = userServiceV1.findUser(userId);
         return ResponseEntity.ok(ApiResponse.success(userResponse, "회원 조회 완료"));
     }
 
-    @Operation(summary = "사용자 정보 수정", description = "비밀번호를 제외한 본인의 정보를 수정함")
     @PatchMapping
     public ResponseEntity<ApiResponse<Void>> updateUser(
             @RequestHeader("X-User-Id") Long userId, @Valid @RequestBody UserUpdateRequest userUpdateRequest) {
@@ -37,7 +32,6 @@ public class UserControllerV1 {
         return ResponseEntity.ok(ApiResponse.success("회원 정보 수정 완료"));
     }
 
-    @Operation(summary = "비밀번호 변경", description = "본인의 비밀번호를 변경함")
     @PutMapping("/password")
     public ResponseEntity<ApiResponse<Void>> updateUserPassword(
             @RequestHeader("X-User-Id") Long userId, @Valid @RequestBody UserUpdatePasswordRequest userUpdatePasswordRequest) {
@@ -45,7 +39,6 @@ public class UserControllerV1 {
         return ResponseEntity.ok(ApiResponse.success("비밀번호 수정 완료"));
     }
 
-    @Operation(summary = "회원 탈퇴", description = "비밀번호 입력시 회원 탈퇴")
     @DeleteMapping
     public ResponseEntity<ApiResponse<Void>> deleteUser(
             @RequestHeader("X-User-Id") Long userId, @Valid @RequestBody UserDeleteRequest userDeleteRequest) {
