+++ src/main/resources/static/docs/index.html
@@ -1154,7 +1154,7 @@ Content-Length: 88
 <div id="footer">
 <div id="footer-text">
 Version 0.0.1-SNAPSHOT<br>
-Last updated 2025-04-29 14:43:55 +0900
+Last updated 2025-04-29 15:45:59 +0900
 </div>
 </div>
 </body>
