+++ src/test/java/com/couponmoa/backend/couponmoauser/domain/user/controller/UserProfileControllerTest.java
@@ -1,10 +1,10 @@
-package com.couponmoa.backend.couponmoauser.domain.user.controller.v1;
+package com.couponmoa.backend.couponmoauser.domain.user.controller;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.couponmoa.backend.couponmoauser.common.service.RedisService;
 import com.couponmoa.backend.couponmoauser.config.JwtUtil;
 import com.couponmoa.backend.couponmoauser.config.SecurityConfig;
-import com.couponmoa.backend.couponmoauser.domain.user.service.v1.UserProfileService;
+import com.couponmoa.backend.couponmoauser.domain.user.service.UserProfileService;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
@@ -19,7 +19,6 @@ import org.springframework.web.multipart.MultipartFile;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.willDoNothing;
-import static org.springframework.restdocs.headers.HeaderDocumentation.headerWithName;
 import static org.springframework.restdocs.headers.HeaderDocumentation.requestHeaders;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.request.RequestDocumentation.partWithName;
